+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.6.0
+
+- ItemStack#hasEnchantment
+
 # 1.5.0
 
 - weakMutableSet()